ST-25-refactor: 캘린더 컴포넌트, 로직 분리
