+++ src/app/[activityId]/_components/ReservationSteps/Calendar/index.tsx
@@ -1,0 +1,63 @@
+"use client";
+
+import CalendarDay from "@/app/[activityId]/_components/ReservationSteps/Calendar/CalendarDay";
+import { ScheduleItem, getDaysArray, monthNames } from "@/app/[activityId]/_utils/createCalendar";
+import { useState } from "react";
+
+interface CalendarProps {
+  scheduleData: ScheduleItem[];
+  setSelectedDate: (date: string) => void;
+}
+
+function Calendar({ scheduleData, setSelectedDate }: CalendarProps) {
+  const [currentDate, setCurrentDate] = useState(new Date());
+  const [selectedDay, setSelectedDay] = useState<string | null>(null);
+
+  const changeMonth = (increment: number) => {
+    setCurrentDate(date => new Date(date.getFullYear(), date.getMonth() + increment, 1));
+  };
+
+  const handleDateSelection = (formattedDate: string) => {
+    setSelectedDay(formattedDate);
+    setSelectedDate(formattedDate);
+  };
+
+  return (
+    <div className="mx-auto mt-4 w-80 overflow-hidden rounded-lg border">
+      <div className="flex items-center justify-between px-4 pt-2">
+        <button type="button" onClick={() => changeMonth(-1)}>
+          &laquo;
+        </button>
+        <p className="text-md font-bold">{`${monthNames[currentDate.getMonth()]} ${currentDate.getFullYear()}`}</p>
+        <button type="button" onClick={() => changeMonth(1)}>
+          &raquo;
+        </button>
+      </div>
+      <div className="grid grid-cols-7 gap-[1px] px-7">
+        {["Su", "Mo", "Tu", "We", "Th", "Fr", "Sa"].map(day => (
+          <p
+            key={day}
+            className="flex items-center justify-center p-2 text-md font-bold text-gray-700"
+          >
+            {day}
+          </p>
+        ))}
+        {getDaysArray(currentDate, scheduleData).map(
+          ({ formattedDate, date, isCurrentMonth, hasReservation }) => (
+            <CalendarDay
+              key={date.toISOString()}
+              date={date}
+              formattedDate={formattedDate}
+              isCurrentMonth={isCurrentMonth}
+              hasReservation={hasReservation}
+              isSelected={formattedDate === selectedDay}
+              setSelectedDate={handleDateSelection}
+            />
+          ),
+        )}
+      </div>
+    </div>
+  );
+}
+
+export default Calendar;
