+++ src/app/[activityId]/_components/ReservationSteps/FreeReservationTime.tsx
@@ -5,9 +5,9 @@ function FreeReservationTime({ reservationTimes }: { reservationTimes: string[][
       <div className="mt-3 grid w-fit grid-cols-2 gap-3">
         {reservationTimes.map(item => (
           <button
-            key={item.join("~")}
+            key={crypto.randomUUID()}
             type="button"
-            className="flex h-11 w-28 items-center justify-center rounded-lg border border-primary-400 text-lg font-medium text-primary-400 hover:bg-primary-400 hover:text-white"
+            className="flex h-11 w-28 items-center justify-center rounded-lg border border-primary-400 text-lg font-medium text-primary-400 focus:bg-primary-400 focus:text-white"
           >
             {item.join("~")}
           </button>
